requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
python-multipart>=0.0.12,<1.0
opencv-python-headless>=4.10,<5.0
pymongo>=4.10,<5.0
PyJWT>=2.9,<3.0
